Auto-rename keys when changing a metadata key to an existing name
Previously, updateMetadataFieldKey would refuse to rename keys to an existing name. This leads to confusing behaviour: If you have have a field "A" and a field "B", renaming "A" to "B" appears to create two fields "B". However, `state` still contains "A" and "B" because updateMetadataFieldKey exits early.

To fix this, updateMetadataFieldKey now renames an existing key with as "$EXISTING_KEYNAME old".